backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
httpx==0.28.1
python-dotenv==1.1.1
joblib==1.5.1
scikit-learn==1.7.1
vobject==0.9.9
icalendar==6.3.1
supabase==2.16.0
redis==6.2.0
pydantic==2.11.7
